requirements.txt
# EchoCore 依赖
# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ASR客户端
websockets>=12.0

# LLM集成
aiohttp>=3.9.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 异步文件操作
aiofiles>=23.2.1

# 配置管理
pyyaml>=6.0.1

# 认证
pyjwt>=2.8.0
cryptography>=41.0.0

# 工具库
python-dateutil>=2.8.2
uuid>=1.30

# FunASR 依赖
funasr>=1.3.0
torchaudio>=2.4.0
addict>=2.4.0
modelscope>=1.9.0
humanfriendly>=10.0

# 音频处理（前端用）
# recorder-core.js (已有)
